+++ src/main/java/com/library/libraryapi/DAO/BookRepository.java
@@ -1,0 +1,7 @@
+package com.library.libraryapi.DAO;
+
+import com.library.libraryapi.Model.Book;
+import org.springframework.data.repository.CrudRepository;
+
+public interface BookRepository extends CrudRepository <Book, Long>{
+}