+++ src/main/java/com/library/libraryapi/Model/Book.java
@@ -1,0 +1,124 @@
+package com.library.libraryapi.Model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
+
+@Entity
+public class Book {
+
+    @Id
+    @GeneratedValue
+    @NotNull
+    private long id;
+
+    @NotNull
+    private String isbn;
+
+    @NotNull
+    private String title;
+
+    @NotNull
+    private String author;
+
+    @NotNull
+    private String category;
+
+    @NotNull
+    private String publishing;
+
+    @NotNull
+    private int publishingYear;
+
+    private int pages;
+    private String description;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getIsbn() {
+        return isbn;
+    }
+
+    public void setIsbn(String isbn) {
+        this.isbn = isbn;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(String author) {
+        this.author = author;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    public String getPublishing() {
+        return publishing;
+    }
+
+    public void setPublishing(String publishing) {
+        this.publishing = publishing;
+    }
+
+    public int getPublishingYear() {
+        return publishingYear;
+    }
+
+    public void setPublishingYear(int publishingYear) {
+        this.publishingYear = publishingYear;
+    }
+
+    public int getPages() {
+        return pages;
+    }
+
+    public void setPages(int pages) {
+        this.pages = pages;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return "Book{" +
+                "id=" + id +
+                ", isbn='" + isbn + '\'' +
+                ", title='" + title + '\'' +
+                ", author='" + author + '\'' +
+                ", category='" + category + '\'' +
+                ", publishing='" + publishing + '\'' +
+                ", publishingYear=" + publishingYear +
+                ", pages=" + pages +
+                ", description='" + description + '\'' +
+                '}';
+    }
+}
+